cleanup, current user lookup, etc
