+++ user/schema.py
@@ -8,11 +8,10 @@ if projdir not in sys.path:
     sys.path.append(appdir)
     sys.path.append(projdir)
 
-
 import graphene
 from graphene import relay
 from graphene_sqlalchemy import SQLAlchemyObjectType, SQLAlchemyConnectionField
-from config.project_globals import (DBSession,Base,metadata,engine,get_scoped_session,
+from config.project_globals import (DBSession,Base,metadata,engine,ScopedSession,
     app,api,ScopedSession)
 from config.settings import VERSION
 from auth.auth_api import jwt_required
@@ -21,7 +20,7 @@ from flask_graphql import GraphQLView
 from auth.fernet_crypto import encrypt,decrypt
 from utils import graphql_utils
 from auth.auth_mutation import AuthMutation,RefreshMutation
-
+from user.user_utils import current_user
 
 class UserAttribute:
     username = graphene.String(description="Username")
@@ -57,24 +56,24 @@ class CreateUser(graphene.Mutation):
     user = graphene.Field(lambda: User, description="User created by this mutation.")
 
     class Arguments:
-        input = CreateUserInput(required=True)
+        inbound = CreateUserInput(required=True)
 
-    def mutate(self, info, input):
-        data = graphql_utils.input_to_dictionary(input)
+    def mutate(self, info, inbound):
+        data = graphql_utils.input_to_dictionary(inbound)
 
         user = UserModel(**data)
+        user_id = None
         # Add validation for non-empty passwords, etc.
         user.password = encrypt(user.password)
-        local_db_session = get_scoped_session()
-        local_db_session.add(user)
-        local_db_session.flush()
-        user_id = user.id
-        local_db_session.commit()
+        with ScopedSession as local_db_session:
+            local_db_session.add(user)
+            local_db_session.flush()
+            user_id = user.id
 
         user = DBSession.query(UserModel).filter(UserModel.id==user_id).first()
         return CreateUser(user=user)
 
-class UpdateUserInput(graphene.InputObjectType,UserAttribute,):
+class UpdateUserInput(graphene.InputObjectType,UserAttribute):
     """Arguments to update a user."""
     id = graphene.ID(required=True, description="Global Id of the user.")
 
@@ -83,33 +82,90 @@ class UpdateUser(graphene.Mutation):
     user = graphene.Field(lambda: User, description="User updated by this mutation.")
 
     class Arguments:
-        input = UpdateUserInput(required=True)
+        inbound = UpdateUserInput(required=True)
 
-    # decorate this with jwt login decorator.
-    def mutate(self, info, input):
-        data = graphql_utils.input_to_dictionary(input)
-        local_db_session = get_scoped_session()
-
-        user = local_db_session.query(UserModel)\
-            .filter(UserModel.id==data['id']).first()
-        for key, value in data.items():
-            if hasattr(user, key):
-                setattr(user, key, value)
-        local_db_session.commit()
+    @jwt_required
+    def mutate(self, info, inbound):
+        data = graphql_utils.input_to_dictionary(inbound)
+        with ScopedSession as local_db_session:
+            user = local_db_session.query(UserModel)\
+                .filter(UserModel.id==data['id']).first()
+            for key, value in data.items():
+                if hasattr(user, key):
+                    setattr(user, key, value)
         user = DBSession.query(UserModel).filter(UserModel.id==data['id']).first()
-
         return UpdateUser(user=user)
 
-class CurrentUserInput(graphene.InputObjectType):
-    username = graphene.String(required=False)
-
 class OneUserInput(graphene.InputObjectType):
+    """ This is used to get one user's data, including their global ID.
+        This is why the global ID is not required. 
+        You should be able to pass nothing and get data on your own logged-in user,
+        because of the JWT token. """
+    id = graphene.ID(required=False, description="Global Id of the user.")
     username = graphene.String(required=False)
     db_id = graphene.Int(required=False)
     email = graphene.String(required=False)
 
+class OneUser(graphene.ObjectType):
+    """User lookup. If not an admin, you can only look yourself up,
+       and only if you're logged in."""
+    user = graphene.Field(lambda: User, description="User lookup for current user, or admin other user lookup.")
+
+    class Arguments:
+        """ Looking up current user requires no arguments. """
+        inbound = OneUserInput(required=False)
+
+    @jwt_required
+    def resolve_search(self, info, inbound):
+        """ Get user from JWT token. """
+        code,error,this_user,timezone = current_user()
+        if code != 200:
+            raise Exception(error)
+
+        """ Compare it with params. If current user is an admin, allow lookup
+            of other users. 
+        """
+        if inbound:
+            data = graphql_utils.input_to_dictionary(inbound)
+            lookup_user = DBSession.query(UserModel)
+            if 'id' in data and data['id']:
+                lookup_user = lookup_user.filter(UserModel.id==data['id'])
+            if 'id' in data and data['id']:
+                lookup_user = lookup_user.filter(UserModel.id==data['id'])
+
+        access_token = request.headers.get(app.config['JWT_HEADER_NAME'],None)
+        #app.logger.info("access token:{0}".format(access_token))
+
+        # If latest user session access token doesn't match, kick them out.
+        user_session = DBSession.query(UserSession).filter(
+            UserSession.user_id==user.id).order_by(
+            UserSession.recorded_time.desc()).first()
+
+        if not user_session:
+            raise Exception('Bad user session')
+
+        if (user_session.access_token != access_token):
+            TNL.add(access_token)
+            # No. user session may be valid, from a newer login on a different device.
+            #TNL.add(user_session.refresh_token)
+            #TNL.add(user_session.access_token)
+            raise Exception('Access token is invalid')
+
+        self.result = {
+            'user_id':user.id,'role':user.role,
+            'phone':user.phone,
+            'first_name':user.first_name, 'last_name': user.last_name,
+            'email':user.email,
+            'timezone':timezone if timezone != pytz.UTC else None,
+            'groups':[],
+            'username':user.username,
+            }
+        return graphql_utils.json2obj(self.result)
+
+
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 class Mutation(graphene.ObjectType):
+    """ Some mutations are imported from auth/ """
     createUser = CreateUser.Field()
     updateUser = UpdateUser.Field()
     authUser = AuthMutation.Field()
